Checked

+++ TS_ES_EcommerceSystem/Models/WarehouseModel/ProductionBatch.cs
@@ -6,7 +6,7 @@
         public string? ProductionBatchName { get; set; }
         public int ProductID { get; set; }
         public int Quantity { get; set; }
-        public float PriceOfBatch { get; set; }
+        public double PriceOfBatch { get; set; }
         public int UnitID { get; set; }
         public DateTime ManufactureDate { get; set; }
         public DateTime ExpiryDate { get; set; }
